Fix PCI enumeration

+++ Sample/App/AppLib/PciInfo.h
@@ -11,7 +11,7 @@ typedef struct _PCI_DEV_LIST PCI_DEV_LIST;
 
 typedef struct _PCI_DEV_LIST {
     UINT32   PFA;
-    UINT8    Bus;
+    UINT16   Bus;
     UINT8    Dev;
     UINT8    Fun;
     UINT16   VendorId;
